backend/requirements.txt

fastapi==0.115.0
uvicorn==0.30.6
boto3==1.34.162
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
